Fix bug in NewResourcesSubmissionHost for L0
There was a bug - even though we created new buffer each iteration, we still
had the old one from warmup (which was freed) bound to the command list.
This change create a new cmdlist with fresh buffer with each iteration.

Note: it worked on Linux by luck, we were getting the same address each iteration.
On Windows we accessed GraphicsAllocation which was invalid.