hAdd missing non-optional pGlobalWorkOffset to urEnqueueLaunchKernel
UR Spec requires the global offset to be always provided.
This worked until now for L0 because neither adapter actually
dereferenced that pointer, instead it was just passed directly to L0.

However, that changed recently in
https://github.com/intel/llvm/pull/18242, and now the ur SubmitKernel
tests are failing.

Long-term the solution is to make pGlobalWorkOffset optional, but for
now this patch just adds it to the benchmark.

Signed-off-by: Piotr Balcer <piotr.balcer@intel.com>