fix: R&R - use proper cmdlist type in emulated mode
The `root` cmdlist is always immediate and should not be used for
recording in emulation mode.
For immediate cmdlist NEO tries to capture commands into a graph, which
is unexpected in such mode. For regular cmdlists the capturing is
skipped and commands are directly recorded into cmdlists, which is the
expected scenario in emulation mode.

Related-To: NEO-16017
Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>